requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.0
Django>=3.0.14
django-allauth==0.42.0
django-cloudinary-storage==0.2.3
django-cors-headers==3.4.0
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow>=8.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.1.0
wrapt==1.12.1
